Changed how menus are handled. Reverted Game to autoload. Removed Globals.
Menus are now children of the Game scene, rather than changing scenes altogether.
Prevents having to save and load data while running in the future.

Initially removed the Game scene from autoloading and used Globals to store data.
This ended up being more complex and messy, so the change was reverted and the Game scene is once again autoloaded.

Removed the Globals script as it is no longer needed.

Cleaned up references to Globals.gd.